LPS-178898 Extract ScreenNavigationEntry methods from DispatchDetailsScreenNavigationCategory to DispatchDetailsScreenNavigationEntry and extends DispatchDetailsScreenNavigationCategory in it
